+++ js/smooth-scroll.js
@@ -1,21 +1,19 @@
-/**
- * smooth-scroll v5.3.3
- * Animate scrolling to anchor links, by Chris Ferdinandi.
+/*!
+ * smooth-scroll v10.2.1: Animate scrolling to anchor links
+ * (c) 2016 Chris Ferdinandi
+ * MIT License
  * http://github.com/cferdinandi/smooth-scroll
- * 
- * Free to use under the MIT License.
- * http://gomakethings.com/mit/
  */
 
 (function (root, factory) {
 	if ( typeof define === 'function' && define.amd ) {
-		define('smoothScroll', factory(root));
+		define([], factory(root));
 	} else if ( typeof exports === 'object' ) {
 		module.exports = factory(root);
 	} else {
 		root.smoothScroll = factory(root);
 	}
-})(window || this, function (root) {
+})(typeof global !== 'undefined' ? global : this.window || this.global, (function (root) {
 
 	'use strict';
 
@@ -24,17 +22,17 @@
 	//
 
 	var smoothScroll = {}; // Object for public APIs
-	var supports = !!document.querySelector && !!root.addEventListener; // Feature test
-	var settings, eventTimeout, fixedHeader;
+	var supports = 'querySelector' in document && 'addEventListener' in root; // Feature test
+	var settings, anchor, toggle, fixedHeader, headerHeight, eventTimeout, animationInterval;
 
 	// Default settings
 	var defaults = {
+		selector: '[data-scroll]',
+		selectorHeader: null,
 		speed: 500,
 		easing: 'easeInOutCubic',
 		offset: 0,
-		updateURL: true,
-		callbackBefore: function () {},
-		callbackAfter: function () {}
+		callback: function () {}
 	};
 
 
@@ -43,78 +41,92 @@
 	//
 
 	/**
-	 * A simple forEach() implementation for Arrays, Objects and NodeLists
+	 * Merge two or more objects. Returns a new object.
 	 * @private
-	 * @param {Array|Object|NodeList} collection Collection of items to iterate
-	 * @param {Function} callback Callback function for each iteration
-	 * @param {Array|Object|NodeList} scope Object/NodeList/Array that forEach is iterating over (aka `this`)
+	 * @param {Boolean}  deep     If true, do a deep (or recursive) merge [optional]
+	 * @param {Object}   objects  The objects to merge together
+	 * @returns {Object}          Merged values of defaults and options
 	 */
-	var forEach = function (collection, callback, scope) {
-		if (Object.prototype.toString.call(collection) === '[object Object]') {
-			for (var prop in collection) {
-				if (Object.prototype.hasOwnProperty.call(collection, prop)) {
-					callback.call(scope, collection[prop], prop, collection);
-				}
-			}
-		} else {
-			for (var i = 0, len = collection.length; i < len; i++) {
-				callback.call(scope, collection[i], i, collection);
-			}
-		}
-	};
+	var extend = function () {
 
-	/**
-	 * Merge defaults with user options
-	 * @private
-	 * @param {Object} defaults Default settings
-	 * @param {Object} options User options
-	 * @returns {Object} Merged values of defaults and options
-	 */
-	var extend = function ( defaults, options ) {
+		// Variables
 		var extended = {};
-		forEach(defaults, function (value, prop) {
-			extended[prop] = defaults[prop];
-		});
-		forEach(options, function (value, prop) {
-			extended[prop] = options[prop];
-		});
-		return extended;
-	};
+		var deep = false;
+		var i = 0;
+		var length = arguments.length;
 
-	/**
-	 * Get the closest matching element up the DOM tree
-	 * @param {Element} elem Starting element
-	 * @param {String} selector Selector to match against (class, ID, or data attribute)
-	 * @return {Boolean|Element} Returns false if not match found
-	 */
-	var getClosest = function (elem, selector) {
-		var firstChar = selector.charAt(0);
-		for ( ; elem && elem !== document; elem = elem.parentNode ) {
-			if ( firstChar === '.' ) {
-				if ( elem.classList.contains( selector.substr(1) ) ) {
-					return elem;
-				}
-			} else if ( firstChar === '#' ) {
-				if ( elem.id === selector.substr(1) ) {
-					return elem;
-				}
-			} else if ( firstChar === '[' ) {
-				if ( elem.hasAttribute( selector.substr(1, selector.length - 2) ) ) {
-					return elem;
+		// Check if a deep merge
+		if ( Object.prototype.toString.call( arguments[0] ) === '[object Boolean]' ) {
+			deep = arguments[0];
+			i++;
+		}
+
+		// Merge the object into the extended object
+		var merge = function (obj) {
+			for ( var prop in obj ) {
+				if ( Object.prototype.hasOwnProperty.call( obj, prop ) ) {
+					// If deep merge and property is an object, merge properties
+					if ( deep && Object.prototype.toString.call(obj[prop]) === '[object Object]' ) {
+						extended[prop] = extend( true, extended[prop], obj[prop] );
+					} else {
+						extended[prop] = obj[prop];
+					}
 				}
 			}
+		};
+
+		// Loop through each object and conduct a merge
+		for ( ; i < length; i++ ) {
+			var obj = arguments[i];
+			merge(obj);
 		}
-		return false;
+
+		return extended;
+
 	};
 
 	/**
-	 * Get the height of an element
+	 * Get the height of an element.
 	 * @private
-	 * @param  {Node]} elem The element
-	 * @return {Number}     The element's height
+	 * @param  {Node} elem The element to get the height of
+	 * @return {Number}    The element's height in pixels
 	 */
-	var getHeight = function (elem) {
+	var getHeight = function ( elem ) {
 		return Math.max( elem.scrollHeight, elem.offsetHeight, elem.clientHeight );
+	};
+
+	/**
+	 * Get the closest matching element up the DOM tree.
+	 * @private
+	 * @param  {Element} elem     Starting element
+	 * @param  {String}  selector Selector to match against
+	 * @return {Boolean|Element}  Returns null if not match found
+	 */
+	var getClosest = function ( elem, selector ) {
+
+		// Element.matches() polyfill
+		if (!Element.prototype.matches) {
+			Element.prototype.matches =
+				Element.prototype.matchesSelector ||
+				Element.prototype.mozMatchesSelector ||
+				Element.prototype.msMatchesSelector ||
+				Element.prototype.oMatchesSelector ||
+				Element.prototype.webkitMatchesSelector ||
+				function(s) {
+					var matches = (this.document || this.ownerDocument).querySelectorAll(s),
+						i = matches.length;
+					while (--i >= 0 && matches.item(i) !== this) {}
+					return i > -1;
+				};
+		}
+
+		// Get closest match
+		for ( ; elem && elem !== document; elem = elem.parentNode ) {
+			if ( elem.matches( selector ) ) return elem;
+		}
+
+		return null;
+
 	};
 
 	/**
@@ -125,6 +137,12 @@
 	 * @link https://github.com/mathiasbynens/CSS.escape
 	 */
 	var escapeCharacters = function ( id ) {
+
+		// Remove leading hash
+		if ( id.charAt(0) === '#' ) {
+			id = id.substr(1);
+		}
+
 		var string = String(id);
 		var length = string.length;
 		var index = -1;
@@ -186,7 +204,9 @@
 			result += '\\' + string.charAt(index);
 
 		}
-		return result;
+
+		return '#' + result;
+
 	};
 
 	/**
@@ -230,8 +250,17 @@
 				anchor = anchor.offsetParent;
 			} while (anchor);
 		}
-		location = location - headerHeight - offset;
-		return location >= 0 ? location : 0;
+		location = Math.max(location - headerHeight - offset, 0);
+		return Math.min(location, getDocumentHeight() - getViewportHeight());
+	};
+
+	/**
+	 * Determine the viewport's height
+	 * @private
+	 * @returns {Number}
+	 */
+	var getViewportHeight = function() {
+		return Math.max( document.documentElement.clientHeight, root.innerHeight || 0 );
 	};
 
 	/**
@@ -258,46 +287,66 @@
 	};
 
 	/**
-	 * Update the URL
+	 * Get the height of the fixed header
 	 * @private
-	 * @param {Element} anchor The element to scroll to
-	 * @param {Boolean} url Whether or not to update the URL history
+	 * @param  {Node}   header The header
+	 * @return {Number}        The height of the header
 	 */
-	var updateUrl = function ( anchor, url ) {
-		if ( history.pushState && (url || url === 'true') ) {
-			history.pushState( null, null, [root.location.protocol, '//', root.location.host, root.location.pathname, root.location.search, anchor].join('') );
+	var getHeaderHeight = function ( header ) {
+		return !header ? 0 : ( getHeight( header ) + header.offsetTop );
+	};
+
+	/**
+	 * Bring the anchored element into focus
+	 * @private
+	 */
+	var adjustFocus = function ( anchor, endLocation, isNum ) {
+
+		// Don't run if scrolling to a number on the page
+		if ( isNum ) return;
+
+		// Otherwise, bring anchor element into focus
+		anchor.focus();
+		if ( document.activeElement.id !== anchor.id ) {
+			anchor.setAttribute( 'tabindex', '-1' );
+			anchor.focus();
+			anchor.style.outline = 'none';
 		}
+		root.scrollTo( 0 , endLocation );
+
 	};
 
 	/**
 	 * Start/stop the scrolling animation
 	 * @public
-	 * @param {Element} toggle The element that toggled the scroll event
-	 * @param {Element} anchor The element to scroll to
-	 * @param {Object} options
+	 * @param {Node|Number} anchor  The element or position to scroll to
+	 * @param {Element}     toggle  The element that toggled the scroll event
+	 * @param {Object}      options
 	 */
-	smoothScroll.animateScroll = function ( toggle, anchor, options ) {
+	smoothScroll.animateScroll = function ( anchor, toggle, options ) {
 
 		// Options and overrides
-		var settings = extend( settings || defaults, options || {} );  // Merge user options with defaults
 		var overrides = getDataOptions( toggle ? toggle.getAttribute('data-options') : null );
-		settings = extend( settings, overrides );
-		anchor = '#' + escapeCharacters(anchor.substr(1)); // Escape special characters and leading numbers
+		var animateSettings = extend( settings || defaults, options || {}, overrides ); // Merge user options with defaults
 
 		// Selectors and variables
-		var anchorElem = anchor === '#' ? document.documentElement : document.querySelector(anchor);
+		var isNum = Object.prototype.toString.call( anchor ) === '[object Number]' ? true : false;
+		var anchorElem = isNum || !anchor.tagName ? null : anchor;
+		if ( !isNum && !anchorElem ) return;
 		var startLocation = root.pageYOffset; // Current location on the page
-		if ( !fixedHeader ) { fixedHeader = document.querySelector('[data-scroll-header]'); }  // Get the fixed header if not already set
-		var headerHeight = fixedHeader === null ? 0 : ( getHeight( fixedHeader ) + fixedHeader.offsetTop ); // Get the height of a fixed header if one exists
-		var endLocation = getEndLocation( anchorElem, headerHeight, parseInt(settings.offset, 10) ); // Scroll to location
-		var animationInterval; // interval timer
+		if ( animateSettings.selectorHeader && !fixedHeader ) {
+			// Get the fixed header if not already set
+			fixedHeader = document.querySelector( animateSettings.selectorHeader );
+		}
+		if ( !headerHeight ) {
+			// Get the height of a fixed header if one exists and not already set
+			headerHeight = getHeaderHeight( fixedHeader );
+		}
+		var endLocation = isNum ? anchor : getEndLocation( anchorElem, headerHeight, parseInt(animateSettings.offset, 10) ); // Location to scroll to
 		var distance = endLocation - startLocation; // distance to travel
 		var documentHeight = getDocumentHeight();
 		var timeLapsed = 0;
 		var percentage, position;
-
-		// Update URL
-		updateUrl(anchor, settings.updateURL);
 
 		/**
 		 * Stop the scroll animation when it reaches its target (or the bottom/top of page)
@@ -306,12 +355,19 @@
 		 * @param {Number} endLocation Scroll to location
 		 * @param {Number} animationInterval How much to scroll on this loop
 		 */
-		var stopAnimateScroll = function (position, endLocation, animationInterval) {
+		var stopAnimateScroll = function ( position, endLocation, animationInterval ) {
 			var currentLocation = root.pageYOffset;
 			if ( position == endLocation || currentLocation == endLocation || ( (root.innerHeight + currentLocation) >= documentHeight ) ) {
+
+				// Clear the animation timer
 				clearInterval(animationInterval);
-				anchorElem.focus();
-				settings.callbackAfter( toggle, anchor ); // Run callbacks after animation complete
+
+				// Bring the anchored element into focus
+				adjustFocus( anchor, endLocation, isNum );
+
+				// Run callback after animation complete
+				animateSettings.callback( anchor, toggle );
+
 			}
 		};
 
@@ -321,9 +377,9 @@
 		 */
 		var loopAnimateScroll = function () {
 			timeLapsed += 16;
-			percentage = ( timeLapsed / parseInt(settings.speed, 10) );
+			percentage = ( timeLapsed / parseInt(animateSettings.speed, 10) );
 			percentage = ( percentage > 1 ) ? 1 : percentage;
-			position = startLocation + ( distance * easingPattern(settings.easing, percentage) );
+			position = startLocation + ( distance * easingPattern(animateSettings.easing, percentage) );
 			root.scrollTo( 0, Math.floor(position) );
 			stopAnimateScroll(position, endLocation, animationInterval);
 		};
@@ -333,7 +389,7 @@
 		 * @private
 		 */
 		var startAnimateScroll = function () {
-			settings.callbackBefore( toggle, anchor ); // Run callbacks before animating scroll
+			clearInterval(animationInterval);
 			animationInterval = setInterval(loopAnimateScroll, 16);
 		};
 
@@ -351,15 +407,101 @@
 	};
 
 	/**
+	 * Handle has change event
+	 * @private
+	 */
+	var hashChangeHandler = function (event) {
+
+		// Get hash from URL
+		// var hash = decodeURIComponent( escapeCharacters( root.location.hash ) );
+		var hash;
+		try {
+			hash = escapeCharacters( decodeURIComponent( root.location.hash ) );
+		} catch(e) {
+			hash = escapeCharacters( root.location.hash );
+		}
+
+		// Only run if there's an anchor element to scroll to
+		if ( !anchor ) return;
+
+		// Reset the anchor element's ID
+		anchor.id = anchor.getAttribute( 'data-scroll-id' );
+
+		// Scroll to the anchored content
+		smoothScroll.animateScroll( anchor, toggle );
+
+		// Reset anchor and toggle
+		anchor = null;
+		toggle = null;
+
+	};
+
+	/**
 	 * If smooth scroll element clicked, animate scroll
 	 * @private
 	 */
-	var eventHandler = function (event) {
-		var toggle = getClosest(event.target, '[data-scroll]');
-		if ( toggle && toggle.tagName.toLowerCase() === 'a' ) {
-			event.preventDefault(); // Prevent default click event
-			smoothScroll.animateScroll( toggle, toggle.hash, settings); // Animate scroll
+	var clickHandler = function (event) {
+
+		// Don't run if right-click or command/control + click
+		if ( event.button !== 0 || event.metaKey || event.ctrlKey ) return;
+
+		// Check if a smooth scroll link was clicked
+		toggle = getClosest( event.target, settings.selector );
+		if ( !toggle || toggle.tagName.toLowerCase() !== 'a' ) return;
+
+		// Only run if link is an anchor and points to the current page
+		if ( toggle.hostname !== root.location.hostname || toggle.pathname !== root.location.pathname || !/#/.test(toggle.href) ) return;
+
+		// Get the sanitized hash
+		// var hash = decodeURIComponent( escapeCharacters( toggle.hash ) );
+		// console.log(hash);
+		var hash;
+		try {
+			hash = escapeCharacters( decodeURIComponent( toggle.hash ) );
+		} catch(e) {
+			hash = escapeCharacters( toggle.hash );
 		}
+
+		// If the hash is empty, scroll to the top of the page
+		if ( hash === '#' ) {
+
+			// Prevent default link behavior
+			event.preventDefault();
+
+			// Set the anchored element
+			anchor = document.body;
+
+			// Save or create the ID as a data attribute and remove it (prevents scroll jump)
+			var id = anchor.id ? anchor.id : 'smooth-scroll-top';
+			anchor.setAttribute( 'data-scroll-id', id );
+			anchor.id = '';
+
+			// If no hash change event will happen, fire manually
+			// Otherwise, update the hash
+			if ( root.location.hash.substring(1) === id ) {
+				hashChangeHandler();
+			} else {
+				root.location.hash = id;
+			}
+
+			return;
+
+		}
+
+		// Get the anchored element
+		anchor = document.querySelector( hash );
+
+		// If anchored element exists, save the ID as a data attribute and remove it (prevents scroll jump)
+		if ( !anchor ) return;
+		anchor.setAttribute( 'data-scroll-id', anchor.id );
+		anchor.id = '';
+
+		// If no hash change event will happen, fire manually
+		if ( toggle.hash === root.location.hash ) {
+			event.preventDefault();
+			hashChangeHandler();
+		}
+
 	};
 
 	/**
@@ -368,12 +510,12 @@
 	 * @param  {Function} eventTimeout Timeout function
 	 * @param  {Object} settings
 	 */
-	var eventThrottler = function (event) {
+	var resizeThrottler = function (event) {
 		if ( !eventTimeout ) {
-			eventTimeout = setTimeout(function() {
+			eventTimeout = setTimeout((function() {
 				eventTimeout = null; // Reset timeout
-				headerHeight = fixedHeader === null ? 0 : ( getHeight( fixedHeader ) + fixedHeader.offsetTop ); // Get the height of a fixed header if one exists
-			}, 66);
+				headerHeight = getHeaderHeight( fixedHeader ); // Get the height of a fixed header if one exists
+			}), 66);
 		}
 	};
 
@@ -387,13 +529,17 @@
 		if ( !settings ) return;
 
 		// Remove event listeners
-		document.removeEventListener( 'click', eventHandler, false );
-		root.removeEventListener( 'resize', eventThrottler, false );
+		document.removeEventListener( 'click', clickHandler, false );
+		root.removeEventListener( 'resize', resizeThrottler, false );
 
 		// Reset varaibles
 		settings = null;
-		eventTimeout = null;
+		anchor = null;
+		toggle = null;
 		fixedHeader = null;
+		headerHeight = null;
+		eventTimeout = null;
+		animationInterval = null;
 	};
 
 	/**
@@ -411,11 +557,19 @@
 
 		// Selectors and variables
 		settings = extend( defaults, options || {} ); // Merge user options with defaults
-		fixedHeader = document.querySelector('[data-scroll-header]'); // Get the fixed header
+		fixedHeader = settings.selectorHeader ? document.querySelector( settings.selectorHeader ) : null; // Get the fixed header
+		headerHeight = getHeaderHeight( fixedHeader );
 
 		// When a toggle is clicked, run the click handler
-		document.addEventListener('click', eventHandler, false );
-		if ( fixedHeader ) { root.addEventListener( 'resize', eventThrottler, false ); }
+		document.addEventListener( 'click', clickHandler, false );
+
+		// Listen for hash changes
+		root.addEventListener('hashchange', hashChangeHandler, false);
+
+		// If window is resized and there's a fixed header, recalculate its size
+		if ( fixedHeader ) {
+			root.addEventListener( 'resize', resizeThrottler, false );
+		}
 
 	};
 
@@ -426,4 +580,4 @@
 
 	return smoothScroll;
 
-});
+}));